Remove the FLNK to Start
Further testing seemed to imply that this is not the most appropriate option. Whilst there is still some peculiarity related to this power supply, until there is further access to test removing the FLNK will at least allow the START to be in place so that it can be sent as required.