Check and resend setpoint
If the raw (as in steam device string) setpoint and readback
differ by more than $(P)SP:CHKTOL then a request is made to
reissue the setpoint after 1 second. The delay is to reduce
problems from a long retry loop.

To test, you could set  $(P)SP:CHKTOL to  0  with a CAPUT and
the the next setpoint should get resent every second to the hardware,
change it back to 0.01 and the sending should then stop.