Added DI:CNT template and removed hardware related CNT templates
The MOXA e1210's hardware based counter API does not seem to be active
on the specific model we have (ioLOGIK E1210) so we have implemented a
EPICS based counter using a calcout record DI:CNT that will keep track
of triggers on the ports.

The vendor provided db's have a number of unused records that relate
to the device being in a counter mode. These are not used by us and
were constantly sending error messages so these have been removed.