Allow the DG645 to be controlled via ethernet
This is much easier than having to faff about with hooking up a moxa.

The process of setting up the device is a bit more involved though.

The following settings need to be configured beforehand:

In the NET menu (SHIFT -> STO(NET)):
TCPIP = Enabled
DHCP = Enabled
Telnet = Enabled
Reset = Yes
Press the Enter Button

1. Then access the status menu (SHIFT -> 6(STATUS) and use the arrows to
navigate to TCPIP Status.
2. Press 6 until you get to IP
3. Make a note of this IP
4. Input this IP into the macro within IBEX
5. Start the IOC

This IP will be reassigned if the device is down for a while. So the
above process will need to be repeated.